Update kiểu dữ liệu int -> float max.py

+++ max.py
@@ -1,7 +1,7 @@
 #Nhập giá trị đầu vào 
-A = int(input("nhập giát trị A = "))
-B = int(input("nhập giát trị B = "))
-C = int(input("nhập giát trị C = "))
+A = float(input("nhập giát trị A = "))
+B = float(input("nhập giát trị B = "))
+C = float(input("nhập giát trị C = "))
 
 #giả định gán cho A là giá trị lớn nhất
 Max = A
